refactor(nms): Separate the EnodebContext
